# sdk/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ApproveOrderDto": {
        "properties": {
          "by": {
            "description": "Identity approving the order",
            "example": "human",
            "type": "string"
          },
          "note": {
            "description": "Optional approval note",
            "type": "string"
          }
        },
        "type": "object"
      },
      "CancelOrderDto": {
        "properties": {
          "reason": {
            "description": "Reason for cancellation",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ExecuteOrderDto": {
        "properties": {},
        "type": "object"
      },
...
      "ProposeOrderDto": {
        "properties": {
          "action": {
            "description": "Order action",
            "enum": ["buy", "sell"],
            "type": "string"
          },
          "address": {
            "description": "Token contract address",
            "example": "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
            "type": "string"
          },
          "amount": {
            "description": "Amount to trade (percentage, \"all\", or USD value as string)",
            "example": "100",
            "type": "string"
          },
          "analysis_score": {
            "description": "Analysis score (0–100)",
            "type": "number"
          },
          "chain": {
            "description": "Chain identifier",
            "example": "base",
            "type": "string"
          },
          "entry_price": {
            "description": "Entry price in USD",
            "type": "number"
          },
          "name": {
            "description": "Token name",
            "example": "Ethereum",
            "type": "string"
          },
          "percent_of_portfolio": {
            "description": "Percent of portfolio",
            "example": 5,
            "type": "number"
          },
          "reason": {
            "description": "Sell reason (for sell orders)",
            "type": "string"
          },
          "reasoning": {
            "description": "Reasoning for this order",
            "type": "string"
          },
          "risk_score": {
            "description": "Risk score (0–100)",
            "type": "number"
          },
          "stop_loss": {
            "description": "Stop-loss price",
            "type": "number"
          },
          "symbol": {
            "description": "Token symbol",
            "example": "ETH",
            "type": "string"
          },
          "take_profit_levels": {
            "description": "Take-profit levels (array of numbers)",
            "items": {
              "type": "number"
            },
            "type": "array"
          },
          "tier": {
            "description": "Position tier",
            "example": "conviction",
            "type": "string"
          },
          "urgency": {
            "description": "Sell urgency",
            "example": "immediate",
            "type": "string"
          }
        },
        "required": ["action", "symbol", "address", "chain", "amount"],
        "type": "object"
      },
      "RejectOrderDto": {
        "properties": {
          "reason": {
            "description": "Reason for rejection",
            "type": "string"
          }
        },
        "type": "object"
      },
      "RetryOrderDto": {
        "properties": {
          "by": {
            "description": "Identity initiating the retry",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "CryptoClaw API — auto-generated from controllers + DTOs",
    "title": "CryptoClaw",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/orders": {
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [
          {
            "description": "Filter by order status",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": ["pending", "approved", "rejected", "cancelled", "executed", "failed", "expired"],
              "type": "string"
            }
          },
          {
            "description": "Filter by action",
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "enum": ["buy", "sell"],
              "type": "string"
            }
          },
          {
            "description": "Show only pending orders",
            "in": "query",
            "name": "pending",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Maximum results (default 50, max 200)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 200,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Cursor for pagination (last order id)",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List orders",
        "tags": ["orders"]
      },
      "post": {
        "operationId": "OrdersController_propose",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeOrderDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Order proposed"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Propose a new order",
        "tags": ["orders"]
      }
    },
    "/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getById",
        "parameters": [
          {
            "description": "Order ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order found"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an order by ID",
        "tags": ["orders"]
      }
    },
    "/v1/orders/{id}/approve": {
      "post": {
        "operationId": "OrdersController_approve",
        "parameters": [
          {
            "description": "Order ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveOrderDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Order approved"
          },
          "404": {
            "description": "Order not found"
          },
          "409": {
            "description": "Invalid state transition"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve an order",
        "tags": ["orders"]
      }
    },
    "/v1/orders/{id}/cancel": {
      "post": {
        "operationId": "OrdersController_cancel",
        "parameters": [
          {
            "description": "Order ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Order cancelled"
          },
          "404": {
            "description": "Order not found"
          },
          "409": {
            "description": "Invalid state transition"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel an order",
        "tags": ["orders"]
      }
    },
    "/v1/orders/{id}/execute": {
      "post": {
        "operationId": "OrdersController_execute",
        "parameters": [
          {
            "description": "Order ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteOrderDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Order execution enqueued (real mode) or completed (paper mode)"
          },
          "404": {
            "description": "Order not found"
          },
          "409": {
            "description": "Order not in approved status"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute an approved order (enqueues BullMQ job in real mode; short-circuits in paper mode)",
        "tags": ["orders"]
      }
    },
    "/v1/orders/{id}/reject": {
      "post": {
        "operationId": "OrdersController_reject",
        "parameters": [
          {
            "description": "Order ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectOrderDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Order rejected"
          },
          "404": {
            "description": "Order not found"
          },
          "409": {
            "description": "Invalid state transition"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject an order",
        "tags": ["orders"]
      }
    },
    "/v1/orders/{id}/retry": {
      "post": {
        "operationId": "OrdersController_retry",
        "parameters": [
          {
            "description": "Order ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryOrderDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Order retried"
          },
          "404": {
            "description": "Order not found"
          },
          "409": {
            "description": "Only failed orders can be retried"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retry a failed order",
        "tags": ["orders"]
      }
    },
...
    "/v1/system/audit": {
      "get": {
        "operationId": "AuditController_list",
        "parameters": [
          {
            "description": "Filter by identity (e.g. RESEARCH, EXECUTOR)",
            "in": "query",
            "name": "identity",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by role (agent|dashboard)",
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "enum": ["agent", "dashboard"],
              "type": "string"
            }
          },
          {
            "description": "Filter by HTTP method",
            "in": "query",
            "name": "method",
            "required": false,
            "schema": {
              "enum": ["GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS", "HEAD"],
              "type": "string"
            }
          },
          {
            "description": "Substring match on path",
            "in": "query",
            "name": "pathContains",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by HTTP status code",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "description": "Return entries from this ISO timestamp onward",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Return entries up to this ISO timestamp",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum results (default 100, max 1000)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Keyset cursor (last seen id from previous page)",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated audit entries"
          },
          "400": {
            "description": "Invalid query parameters"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Query audit log entries",
        "tags": ["system"]
      }
    },
    "/v1/system/audit/{id}": {
      "get": {
        "operationId": "AuditController_getById",
        "parameters": [
          {
            "description": "Audit entry ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit entry found"
          },
          "404": {
            "description": "Audit entry not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a single audit entry by ID",
        "tags": ["system"]
      }
    }
  },
  "servers": [],
...
}